Add basic integration tests
Moved QEMU-specific code to its own lib so that both main and
integration test binaries can use it.
