+++ tests/basic_boot.rs
@@ -1,0 +1,27 @@
+#![no_std]
+#![no_main]
+#![feature(custom_test_frameworks)]
+#![test_runner(rustos::test_runner)]
+#![reexport_test_harness_main = "test_main"]
+
+use core::panic::PanicInfo;
+use rustos::{println, serial_print, serial_println};
+
+#[no_mangle]
+pub extern "C" fn _start() -> ! {
+    test_main();
+
+    loop {}
+}
+
+#[panic_handler]
+fn panic(info: &PanicInfo) -> ! {
+    rustos::test_panic_handler(info)
+}
+
+#[test_case]
+fn test_println() {
+    serial_print!("basic_boot: test_println... ");
+    println!("test_println output");
+    serial_println!("[ok]");
+}
